tools/colour_testbed/requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
pyserial>=3.5

